==> 3_preprocess/2_data_preparation_for_FL.ipynb <==
import pandas as pd
from pathlib import Path
import pickle

PROJECT_DIR =  Path.cwd().parent
DATA_DIR = PROJECT_DIR.joinpath('data')

HICU = pd.read_feather(DATA_DIR.joinpath('processed','HICU.feather'))
SICU = pd.read_feather(DATA_DIR.joinpath('processed','SICU.feather'))
MICU = pd.read_feather(DATA_DIR.joinpath('processed','MICU.feather'))
NSICU = pd.read_feather(DATA_DIR.joinpath('processed','NSICU.feather'))

from sklearn.preprocessing import MinMaxScaler

scaler = MinMaxScaler()

output = scaler.fit_transform(HICU[HICU.columns[1:]])

HICU[HICU.columns[1:]] = output
HICU.head()

HICU

output = scaler.fit_transform(MICU[MICU.columns[1:]])

MICU[MICU.columns[1:]] = output
MICU.head()

MICU

output = scaler.fit_transform(SICU[SICU.columns[1:]])

SICU[SICU.columns[1:]] = output
SICU.head()

output = scaler.fit_transform(NSICU[NSICU.columns[1:]])

NSICU[NSICU.columns[1:]] = output
NSICU.head()

DATA_DIR.joinpath('processed','HICU.csv)

SICU

DATA_DIR.joinpath('processed','HICU.csv')

HICU.to_csv(DATA_DIR.joinpath('processed','HICU.csv'), index=False)
SICU.to_csv(DATA_DIR.joinpath('processed','SICU.csv'), index=False)
MICU.to_csv(DATA_DIR.joinpath('processed','MICU.csv'), index=False)
NSICU.to_csv(DATA_DIR.joinpath('processed','NSICU.csv'), index=False)

HICU.to_csv(PROJECT_DIR.joinpath('4_ppfl_simulation/hicu/data/HICU.csv'), index=False)
SICU.to_csv(PROJECT_DIR.joinpath('4_ppfl_simulation/sicu/data/SICU.csv'), index=False)
MICU.to_csv(PROJECT_DIR.joinpath('4_ppfl_simulation/micu/data/MICU.csv'), index=False)
NSICU.to_csv(PROJECT_DIR.joinpath('4_ppfl_simulation/NSicu/data/NSICU.csv'), index=False)

load valid data

with open(DATA_DIR.joinpath('processed','micu_selected.pkl'),'rb') as f:
   micu_selected = pickle.load(f)
with open(DATA_DIR.joinpath('processed','sicu_selected.pkl'), 'rb') as f:
   sicu_selected = pickle.load(f)
with open(DATA_DIR.joinpath('processed','nsicu_selected.pkl'), 'rb') as f:
   nsicu_selected = pickle.load(f)

common_features = set(MICU.columns) & set(SICU.columns) & set(HICU.columns) & set(MICU.columns)
id_and_outcome = set(['연구등록번호','death'])

micu_specific = set(MICU.columns) - common_features
sicu_specific = set(SICU.columns) - common_features
hicu_specific = set(HICU.columns) - common_features

common_features = common_features - id_and_outcome

feature_book = {'common_features':list(common_features), 'id_and_outcome':list(id_and_outcome),
                'hicu_specific':list(hicu_specific), 'sicu_specific':list(sicu_selected), 'micu_specific':list(micu_selected), 'nsicu_specific':list(nsicu_selected)}

import yaml

with open(DATA_DIR.joinpath('processed','feature_book.yaml'),'w') as f:
    yaml.dump(feature_book, f, allow_unicode=True)
